refactor: Update dependency entries at a higher level
This will allow us to work with more information so we can do
"append-only" updates.
